add get orders by name query in application layer

+++ Src/Services/Ordering/Ordering.Application/Dtos/OrderItemDto.cs
@@ -4,4 +4,4 @@ public record OrderItemDto(
     Guid OrderId,
     Guid ProductId,
     int Quantity,
-    int Price);
+    decimal Price);
